Changes in replies/topics are duplicated in posts
This is temporary code that allows refactoring in small steps.
That way the existing posts/topics functionality can be ported
step by step to posts. When they finally become empty, they
can be removed (or made subclasses).

Reply and Topic might be a bit too simple to require this,
but hey – this is fun practice.