Add solutions.points
We don't want to calculate points anymore, but store them in the database
instead. This allows us to publish tasks that are not runnable (essays,
notes, etc.). It also simplifies the logic for calculating all the points
of a student, since all the points now are directly stored as numbers in
the database.