Add line numbers to solutions
Fixes #11, although in a rather opinionated way.

First, we're using CodeRay instead of a JavaScript highlighter.
That works a bit better, especially in this case. It's also a
lot more reliable, although not as fast (at least on the server
side of things).

Second, we're "customizing" the output from CodeRay to fit it
in a table that behaves the way we want it to. It is not immediately
obvious why this is necessary - we want to have per-line commenting
and that way it would be easier to both annotate the lines and provide
and interface for commenting. Unfortunately, line numbers get in the
way of selecting and copying the code.

We did this together with @nb. (And I also want to test the new
GitHub notifications on him).