Add more photo sizes
The new sizes are: 150x150, 80x80, 50x50 and 30x30. Not all are used,
but probably all would be useful.

Another important change is that all of them are now named consistently,
thus no need to select the user in order to predict the filename of
their photo. This is a breaking change and it requires regenerating all
versions. This can be done with:

    rake users:photos:recreate

This should be executed after the code is deployed.