Remove default table style
We don't need that. It is tricky, since we adding a non-blue table to
the site requires gymnastics. An example is the styling for CodeRay
snippets.

Instead, all tables in the content mixin are colored correctly. Also, if
we need a table to have that color and want to set it up cheaply, we
just use table.data-table.