Reset all passwords when fetching the database
Since the server has a unique salt, we cannot use the passwords on
production without fetching the salt too. As a workaroud, we can reset
all passwords.

I've added a new rake task, rake dev:fetch, to fetch the database and
invoke it.

I'm not really happy about rake invoking capistrano. This is workflow
and workflow in Rake (for Rails) never sat right with me. For this
reason, let's contain all workflow-related rake tasks in the dev
namespace.