Fix SimpleCov results on some configurations
It's a tricky bug, but on my Linux desktop the results from Cucumber and
RSpec did not get merged. The reason turned out to be that SimpleCov's
guess for which command was ran was failing for some unknown to me RVM
reason and thought both runs are RSpec (thus, overriding the previous
result).

It appears to be fixed when command names are indicated explicitly.