Start lectures generation in a clean environment
This is a classic "bundler within bundler" scenario. The outer one overrides
the Gemfile path of any "inner" `bundle` calls.

The reason for this is that exec'd processes inherit the ENV of their executors.
In our case, the executor is a Ruby process with an already initialized Bundler
ENV. This ENV contains the variable BUNDLE_GEMFILE and it makes other `bundle`
calls (which *may* get executed deep in the process chain that this task triggers),
will also inherit this variable and will disregard their eventual local Gemfiles.

Here's the obligatory SO answer link: http://stackoverflow.com/a/10275315/75715