Reset GEM_HOME before running script/lectures
The exec call in this task keeps biting us. I'd say exec is evil. It inherits
a ton of shit from the current process' ENV and this shit messes up the exec'd
process.

For example, here we have a GEM_HOME set to this project's gem bundle, which
is somewhere in the Capistrano shared path. However, when we want to build the
lectures, we don't want to keep gems in this bundle. Code annotation in the
lectures repo is broken due to this, for example.