Just add a team page
We tried to separate the user model into a student and a teacher, but
there are no explicit benefits we found at the moment. Let's just have
the team page for now and if we feel that at some point, the separation
will make the rest of the code better, we can do it then.