Add a missing teams/show view
And once again, the day is saved! Thanks to the Git Reflog!

The view was missing because after some discussion, all changes
from the team-page branch were discarded, the branch was force-
pushed and that view was lost amidst the big "delete code" fuss.

Shame on us! @gsamokovarov @skanev @mitio Shame, shame, shame!