Don't check Go code with gofmt
There seems to be a copy-pasta issue. When people paste the code in a
textarea and submit, tabs are transformed into spaces. It's not clear at
which point this happens, but it seems like it does.

For now, better to disable `gofmt` processing. Better to have the syntax
check only instead of no checks at all.