Don't depend on CWD when running tests
Tests are run in a Sidekiq thread. Sometimes a broken job may change the
CWD of the main Sidekiq process thus chainging it for all future jobs.
It's better that the test job doesn't depend on the CWD of the process
which is running the tests.

For some language runners, this might result in the following error
visible in the Sideliq queue:

A JSON text must at least contain two octets!