Synchronize output to log file in Ruby task runner
Redirecting stdout and stderr separately to a single file results in
unsynchronized writes to the file. Running a command which writes to
stdout and stderr results in the two outputs "merging" mid-line and
sometimes mid-utf sequence.

If one of the outputs is inserted in the middle of a utf character
sequence then we get `ArgumentError: invalid byte sequence` when trying
to split the JSON output from the log at `Language::Ruby#run_tests`.

The problem is caused by `system` opening the same file two times
(one for stdout and one for stderr) and writing to it at the same time.

The fix specifies the file redirection of both streams at the same time
so that the file is open only once.