Show diffs for revisions and preserve comments for unchanged code
It is easier to see what changes are made in a revision when you are
presented with a diff instead of manually trying to figure out the
differences.

Inline comments made for a revision will be visible in the solution code
at the top of the page if the commented code is not changed. If the line
on which the comment is created changes, the comment will only be
visible in the diff for the revision it was created on.

Inline comments can also be made on the diff. The diffs (revisions) will
not show older comments, even if no changes were made to them.