Fix escaping issue with deleted lines in the diff
The code highlighter escapes all HTML special characters and generates
new HTML that should not be escaped. The DiffLine class assumes the code
line to already be escaped and marks it as html_safe so that the
highlighting is not escaped when rendering.

In the case of diffs, the removed code line is not passed through the
highlighter. This means that the code itself is not escaped when it goes
to CodeLine, which marks it as html_safe anyway.

The issue is fixed by making sure the Diff class always escapes code
lines, even if they don't pass through the highlighter.