Upgrade to rails 4.1, limit rake to < 11.0
On my machine, with postgresql 9.6, the `structure.sql` dump doesn't
seem to work -- the `pg_dump` command has changed its flags.

Upgrading to rails 4.1 fixes the issue for me. With that, rake is also
updated to version 12, which results in a different error, which is why
it's locked to an earlier version.