Add a script to anonymize data of users
This allows us to be GDPR-compliant.

User's personal data is anonymized rather than deleted to avoid
dependency issues. User generated tips, solutions and solution check
logs are left intact as they're not considered personal identifiable
information and in practice would not allow inferring the identity of
the original author while at the same time it will be beneficial to keep
the code and improvement tips provided by course staff.