Don't show first diff
Seems like Rouge is super slow particularly through highlighting the
initial diff. This is surprising, but it's also easily avoidable. The
first diff is extremely uninteresting and usually gets in the way, since
it's all additions.

This change makes it clear the first-rendered diff is, in fact, first
historically and removes the unnecessary "view changes" on it, reducing
noise in the history panel. As a pleasant side-effect, it also massively
improves page performance.