Make the project installable on Ruby 2.5.x
This Ruby, although now EOL, still installs and runs on Ubuntu 20.04
LTS as it can compile against OpenSSL 1.1.x. Ubuntu 20.04 does not ship
older OpenSSL versions by default.

The PG gem is pinned to 0.x because of a dependency in Rails.

The .ruby-version file is now ignored from VCS to allow flexibility in
choosing the specific Ruby version against which the project runs.