Remove exception notifications over email
Since these sites are no longer in active development and are fairly
stable, these notifications are more noise than anything useful.
Exceptions usually occur while folks attempt to exploit the site with
sending broken/erroneous payloads, which is not particularly
interesting.

Exceptions will continue to be logged to the production.log file and can
be examined on demand.

Should this functionality be needed, it must be introduced
conditionally, configured in the per-course-edition files.